+++ src/sdk/getexcelfrompdfs.ts
@@ -72,11 +72,10 @@ export class GetExcelFromPDFs {
       switch (true) {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
-            res.getCsvExtraction200ApplicationJSONObject =
-              utils.deserializeJSONResponse(
-                httpRes?.data,
-                operations.GetCsvExtraction200ApplicationJSON
-              );
+            res.getCsvExtraction200ApplicationJSONObject = utils.objectToClass(
+              httpRes?.data,
+              operations.GetCsvExtraction200ApplicationJSON
+            );
           }
           break;
         case httpRes?.status == 400:
@@ -151,7 +150,7 @@ export class GetExcelFromPDFs {
         case httpRes?.status == 200:
           if (utils.matchContentType(contentType, `application/json`)) {
             res.getExcelExtraction200ApplicationJSONObject =
-              utils.deserializeJSONResponse(
+              utils.objectToClass(
                 httpRes?.data,
                 operations.GetExcelExtraction200ApplicationJSON
               );